Update Fontello package and remove unused SVG icons.
