--- ui/src/components/Buttons/CloseButton.md
@@ -1,4 +1,0 @@
-Example:
-```js
-<CloseButton />
-```
