--- ui/src/components/Buttons/OpenWindowButton.css
@@ -1,11 +1,0 @@
-/**
- * Foundation for Sites by ZURB
- * Version 6.5.3
- * foundation.zurb.com
- * Licensed under MIT Open Source
- */
-@import url("https://fonts.googleapis.com/css?family=Montserrat:400,600,800");
-@import url("https://use.fontawesome.com/releases/v5.8.1/css/all.css");
-.OpenWindowButton i {
-  color: #000;
-  font-size: 14px; }
