--- ui/src/components/Buttons/OpenWindowButton.md
@@ -1,5 +1,0 @@
-OpenWindowButton example:
-
-```js
-<OpenWindowButton href="#" />
-```
