--- ui/src/components/Buttons/OpenWindowButton.scss
@@ -1,6 +1,0 @@
-@import "../../theme/base";
-
-.OpenWindowButton i {
-  color: $black;
-  font-size: 14px;
-}
