--- ui/src/components/Buttons/ViewInfoButton.css
@@ -1,16 +1,0 @@
-/**
- * Foundation for Sites by ZURB
- * Version 6.5.3
- * foundation.zurb.com
- * Licensed under MIT Open Source
- */
-@import url("https://fonts.googleapis.com/css?family=Montserrat:400,600,800");
-@import url("https://use.fontawesome.com/releases/v5.8.1/css/all.css");
-.ViewInfoButton {
-  font-size: 14px;
-  font-weight: 800;
-  letter-spacing: 2px;
-  text-transform: uppercase; }
-  .ViewInfoButton:hover, .ViewInfoButton.active-true {
-    background: #292F36;
-    color: #fffffe; }
