--- ui/src/components/Buttons/ViewInfoButton.md
@@ -1,4 +1,0 @@
-Example:
-```js
-<ViewInfoButton />
-```
