--- ui/src/components/Buttons/ViewInfoButton.scss
@@ -1,11 +1,0 @@
-@import "../../theme/base";
-
-.ViewInfoButton {
-  @include font-bold;
-  // @include border-button;
-
-  &:hover, &.active-true, {
-    background: $charcoal;
-    color: $white;
-  }
-}
