+++ ui/src/components/ComponentInfo.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import FetchingCodeBlock from './FetchingCodeBlock';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import Button from './Buttons/Button'
+import Button from './Buttons'
 import './ComponentInfo.css';
 
 const ComponentInfo = ({ component, sample, used, className, controls }) => {
