--- ui/src/components/Icons/CloseArrow.js
@@ -1,8 +1,0 @@
-import React from 'react';
-// import './Icon.css';
-
-export default () => (
-  <svg className="Icon CloseArrow" x="0px" y="0px" viewBox="0 0 16 16">
-    <path d="M8,0L6.6,1.4L12.2,7H0v2h12.2l-5.6,5.6L8,16l8-8L8,0z" />
-  </svg>
-);
