--- ui/src/components/Icons/OpenNew.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <i className="fas fa-external-link-alt"></i>
-);
