--- ui/src/components/Icons/OpenNew.md
@@ -1,6 +1,0 @@
-Example:
-```js
-<div style={{background: 'black', textAlign: 'center', padding: '1em'}}>
-    <OpenNew />
-</div>
-```
