+++ ui/src/components/NavDrawer.scss
@@ -61,10 +61,9 @@
   }
   i {
     position: absolute;
-    top: 50%;
+    top: 25%;
     left: rem-calc(15);
     font-size: rem-calc(13);
-    margin-top: -.5em;
     color: $white;
   }
 }
