+++ ui/src/fontello/LICENSE.txt
