+++ ui/src/fontello/config.json
@@ -37,6 +37,12 @@
       "src": "fontawesome"
     },
     {
+      "uid": "9dd9e835aebe1060ba7190ad2b2ed951",
+      "css": "search",
+      "code": 59395,
+      "src": "fontawesome"
+    },
+    {
       "uid": "d1e828c2c23a6baf620f3175bfab47cf",
       "css": "menu",
       "code": 59394,
