+++ ui/src/fontello/css/animation.css
