+++ ui/src/fontello/css/fontello-codes.css
@@ -2,5 +2,6 @@
 .icon-close:before { content: '\e800'; } /* '' */
 .icon-expand:before { content: '\e801'; } /* '' */
 .icon-collapse:before { content: '\e802'; } /* '' */
+.icon-search:before { content: '\e803'; } /* '' */
 .icon-new-window:before { content: '\f08e'; } /* '' */
 .icon-menu:before { content: '\f0c9'; } /* '' */
