+++ ui/src/fontello/css/fontello.css
@@ -1,11 +1,11 @@
 @font-face {
   font-family: 'fontello';
-  src: url('../font/fontello.eot?80524244');
-  src: url('../font/fontello.eot?80524244#iefix') format('embedded-opentype'),
-       url('../font/fontello.woff2?80524244') format('woff2'),
-       url('../font/fontello.woff?80524244') format('woff'),
-       url('../font/fontello.ttf?80524244') format('truetype'),
-       url('../font/fontello.svg?80524244#fontello') format('svg');
+  src: url('../font/fontello.eot?59602837');
+  src: url('../font/fontello.eot?59602837#iefix') format('embedded-opentype'),
+       url('../font/fontello.woff2?59602837') format('woff2'),
+       url('../font/fontello.woff?59602837') format('woff'),
+       url('../font/fontello.ttf?59602837') format('truetype'),
+       url('../font/fontello.svg?59602837#fontello') format('svg');
   font-weight: normal;
   font-style: normal;
 }
@@ -15,7 +15,7 @@
 @media screen and (-webkit-min-device-pixel-ratio:0) {
   @font-face {
     font-family: 'fontello';
-    src: url('../font/fontello.svg?80524244#fontello') format('svg');
+    src: url('../font/fontello.svg?59602837#fontello') format('svg');
   }
 }
 */
@@ -58,5 +58,6 @@
 .icon-close:before { content: '\e800'; } /* '' */
 .icon-expand:before { content: '\e801'; } /* '' */
 .icon-collapse:before { content: '\e802'; } /* '' */
+.icon-search:before { content: '\e803'; } /* '' */
 .icon-new-window:before { content: '\f08e'; } /* '' */
 .icon-menu:before { content: '\f0c9'; } /* '' */
