+++ ui/src/fontello/demo.html
@@ -229,11 +229,11 @@ body {
 }
 @font-face {
       font-family: 'fontello';
-      src: url('./font/fontello.eot?2458697');
-      src: url('./font/fontello.eot?2458697#iefix') format('embedded-opentype'),
-           url('./font/fontello.woff?2458697') format('woff'),
-           url('./font/fontello.ttf?2458697') format('truetype'),
-           url('./font/fontello.svg?2458697#fontello') format('svg');
+      src: url('./font/fontello.eot?53950484');
+      src: url('./font/fontello.eot?53950484#iefix') format('embedded-opentype'),
+           url('./font/fontello.woff?53950484') format('woff'),
+           url('./font/fontello.ttf?53950484') format('truetype'),
+           url('./font/fontello.svg?53950484#fontello') format('svg');
       font-weight: normal;
       font-style: normal;
     }
@@ -301,9 +301,10 @@ body {
         <div class="the-icons span3" title="Code: 0xe800"><i class="demo-icon icon-close">&#xe800;</i> <span class="i-name">icon-close</span><span class="i-code">0xe800</span></div>
         <div class="the-icons span3" title="Code: 0xe801"><i class="demo-icon icon-expand">&#xe801;</i> <span class="i-name">icon-expand</span><span class="i-code">0xe801</span></div>
         <div class="the-icons span3" title="Code: 0xe802"><i class="demo-icon icon-collapse">&#xe802;</i> <span class="i-name">icon-collapse</span><span class="i-code">0xe802</span></div>
-        <div class="the-icons span3" title="Code: 0xf08e"><i class="demo-icon icon-new-window">&#xf08e;</i> <span class="i-name">icon-new-window</span><span class="i-code">0xf08e</span></div>
+        <div class="the-icons span3" title="Code: 0xe803"><i class="demo-icon icon-search">&#xe803;</i> <span class="i-name">icon-search</span><span class="i-code">0xe803</span></div>
       </div>
       <div class="row">
+        <div class="the-icons span3" title="Code: 0xf08e"><i class="demo-icon icon-new-window">&#xf08e;</i> <span class="i-name">icon-new-window</span><span class="i-code">0xf08e</span></div>
         <div class="the-icons span3" title="Code: 0xf0c9"><i class="demo-icon icon-menu">&#xf0c9;</i> <span class="i-name">icon-menu</span><span class="i-code">0xf0c9</span></div>
       </div>
     </div>
